requirements.txt
google-adk
litellm
fastapi
uvicorn
httpx
pydantic
openai
streamlit
flask
selenium